Extract TypeNames and Name into their own files
